# src/main/resources/static/swagger.yml
swagger: "2.0"
info:
  description: "Details about the `employee` REST API"
  version: "1.0.0"
  title: "Employee REST Service"
tags:
  - name: "employees"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /api/v1/employees:
    post:
      tags:
        - "employees"
      summary: "Adds an active employee"
      parameters:
      - in: "body"
        name: "body"
        description: "Employee object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/EmployeeRequest"
      responses:
        201:
          description: "Employee has been created successfully"
          schema:
            $ref: "#/definitions/EmployeeResponse"

    get:
      tags:
        - "employees"
      summary: "Finds all active employees"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EmployeeResponse"

  /api/v1/employees/{id}:
    get:
      tags:
        - "employees"
      summary: "Finds an active employee by its id"
      parameters:
      - name: "id"
        in: "path"
        description: "Id of employee to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Employee found"
          schema:
            $ref: "#/definitions/EmployeeResponse"
        404:
          description: "Employee not found"

    put:
      tags:
        - "employees"
      summary: "Updates an existing employee"
      parameters:
      - name: "id"
        in: "path"
        description: "Employee id to update"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Employee object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/EmployeeRequest"
      responses:
        204:
          description: "Employee has been updated successfully"
        404:
          description: "Employee not found"
        405:
          description: "Validation exception"

    delete:
      tags:
      - "employees"
      summary: "Deletes an active employee"
      parameters:
      - name: "api_key"
        in: "header"
        description: "Authentication"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        description: "Employee id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Employee has been deleted successfully"
        401:
          description: "User is not authenticated"
          schema:
            $ref: '#/definitions/UnauthorizedError'
        404:
          description: "Employee not found"
        405:
          description: "Validation exception"
      security:
        - basicAuth: []

securityDefinitions:
  basicAuth:
    type: basic

definitions:
  EmployeeRequest:
    type: "object"
    required:
      - "firstName"
      - "lastName"
      - "dateOfBirth"
      - "dateOfEmployment"
    properties:
      firstName:
        type: "string"
        maxLength: 50
      middleInitial:
        type: "string"
        maxLength: 1
      lastName:
        type: "string"
        maxLength: 50
      dateOfBirth:
        type: "string"
        format: "date"
      dateOfEmployment:
        type: "string"
        format: "date"

  EmployeeResponse:
    type: "object"
    required:
      - "id"
      - "firstName"
      - "lastName"
      - "dateOfBirth"
      - "dateOfEmployment"
      - "status"
    properties:
      id:
        type: "integer"
        format: "int64"
      firstName:
        type: "string"
        maxLength: 50
      middleInitial:
        type: "string"
        maxLength: 1
      lastName:
        type: "string"
        maxLength: 50
      dateOfBirth:
        type: "string"
        format: "date"
      dateOfEmployment:
        type: "string"
        format: "date"
      status:
        type: "string"
        description: "Employment status"
        enum:
        - "ACTIVE"
        - "INACTIVE"

  UnauthorizedError:
    description: "User is not authenticated"